requirements.txt
# Core dependencies
python-dotenv
psycopg2-binary